refactor: 1차 PR 반영 - 역할에 맞게 class를 알맞은 package로 이동, Lombok 활용 등
