+++ BE/src/main/java/codesquad/backend/issuetracker/oauth/application/GithubOAuthClient.java
@@ -1,7 +1,8 @@
-package codesquad.backend.issuetracker.oauth.domain.github;
+package codesquad.backend.issuetracker.oauth.application;
 
+import codesquad.backend.issuetracker.oauth.presentation.dto.GithubToken;
+import codesquad.backend.issuetracker.oauth.presentation.dto.GithubUser;
 import java.net.URI;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
