+++ BE/src/main/java/codesquad/backend/issuetracker/oauth/application/OAuthService.java
@@ -3,31 +3,34 @@ package codesquad.backend.issuetracker.oauth.application;
 import codesquad.backend.issuetracker.user.domain.User;
 import codesquad.backend.issuetracker.user.infrastructure.UserRepository;
 import java.util.Optional;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
+@RequiredArgsConstructor
 @Service
 public class OAuthService {
 
 	private final UserRepository userRepository;
 
-	public OAuthService(UserRepository userRepository) {
-		this.userRepository = userRepository;
-	}
-
 	@Transactional
 	public User upsertUser(User user) {
 		log.debug("Auth ID = {}", user.getAuthId());
 
-		Optional<User> optionalUser = userRepository.findByAuthId(user.getAuthId());
+		User findUser = userRepository.findByAuthId(user.getAuthId())
+			.orElse(user);
+		findUser.update(user);
+		userRepository.save(findUser);
 
-		if (optionalUser.isPresent()) {
-			User findUser = optionalUser.get();
-			user = findUser.update(user);
-		}
-		userRepository.save(user);
+//		Optional<User> optionalUser = userRepository.findByAuthId(user.getAuthId());
+//
+//		if (optionalUser.isPresent()) {
+//			User findUser = optionalUser.get();
+//			user = findUser.update(user);
+//		}
+//		userRepository.save(user);
 		return user;
 	}
 }
