+++ BE/src/test/java/codesquad/backend/issuetracker/oauth/JwtTest.java
@@ -1,7 +1,7 @@
 package codesquad.backend.issuetracker.oauth;
 
 import static org.assertj.core.api.Assertions.*;
-import codesquad.backend.issuetracker.oauth.domain.JwtFactory;
+import codesquad.backend.issuetracker.oauth.application.JwtFactory;
 import codesquad.backend.issuetracker.user.domain.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
